requirements.txt
Flask==2.0.1
pyodbc==4.0.30
python-dotenv==0.19.0
gunicorn==20.1.0
flask-cors==3.0.10
